# packages/api/typescript/public/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "v1",
    "title": "codedm-backend"
  },
  "paths": {
...
    "/v1/threads/{threadId}/issues/{issueId}/operator-questions": {
      "post": {
        "tags": [
          "agent",
          "mcp:issue-handling"
        ],
        "description": "Ask the operator a question (fire-and-forget; surfaces as a Needs-you stop)",
        "operationId": "AskOperator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "question"
                ]
              },
              "examples": {
                "default": {
                  "value": {
                    "question": "Should the refund be full or partial for orders older than 90 days?"
                  },
                  "summary": "Example"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "019e4d24-6524-7041-9e1c-8108180cddae"
          },
          {
            "in": "path",
            "name": "issueId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "019e4d24-6524-7041-9e1c-8108180cddaf"
          }
        ],
        "responses": {
          "200": {
            "description": "Ask the operator a question (fire-and-forget; surfaces as a Needs-you stop)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "delivered": {
                      "type": "boolean",
                      "const": true
                    },
                    "stopId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": [
                    "delivered",
                    "stopId"
                  ],
                  "additionalProperties": false,
                  "examples": [
                    {
                      "delivered": true,
                      "stopId": "019e4d24-6524-7041-9e1c-8108180cdda0"
                    }
                  ]
                },
                "examples": {
                  "default": {
                    "value": {
                      "delivered": true,
                      "stopId": "019e4d24-6524-7041-9e1c-8108180cdda0"
                    },
                    "summary": "Example"
                  }
                }
              }
            }
          }
        },
        "x-mcp-scope": [
          "issue-handling"
        ]
      }
    },
...
    "/v1/threads/{threadId}/issues/{issueId}/stops": {
      "post": {
        "tags": [
          "agent",
          "mcp:issue-handling"
        ],
        "description": "Declare that the agent is blocked and needs the human (approval, classification, …)",
        "operationId": "RaiseStop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "$ref": "#/components/schemas/StopKind"
                  },
                  "detail": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  }
                },
                "required": [
                  "kind",
                  "detail"
                ]
              },
              "examples": {
                "default": {
                  "value": {
                    "kind": "APPROVAL_NEEDED",
                    "detail": "This deletes 3 production tables — confirm before I run it."
                  },
                  "summary": "Example"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "019e4d24-6524-7041-9e1c-8108180cddae"
          },
          {
            "in": "path",
            "name": "issueId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "019e4d24-6524-7041-9e1c-8108180cddaf"
          }
        ],
        "responses": {
          "200": {
            "description": "Declare that the agent is blocked and needs the human (approval, classification, …)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stopId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    }
                  },
                  "required": [
                    "stopId"
                  ],
                  "additionalProperties": false,
                  "examples": [
                    {
                      "stopId": "019e4d24-6524-7041-9e1c-8108180cdda0"
                    }
                  ]
                },
                "examples": {
                  "default": {
                    "value": {
                      "stopId": "019e4d24-6524-7041-9e1c-8108180cdda0"
                    },
                    "summary": "Example"
                  }
                }
              }
            }
          }
        },
        "x-mcp-scope": [
          "issue-handling"
        ]
      }
    },
...
    "/v1/threads/{threadId}/issues/{issueId}/status": {
      "post": {
        "tags": [
          "agent",
          "mcp:issue-handling"
        ],
        "description": "Declare the lifecycle status of an issue (done / needs input)",
        "operationId": "TransitionIssueStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/IssueStatus"
                  },
                  "summary": {
                    "default": "",
                    "type": "string",
                    "maxLength": 4000
                  },
                  "key": {
                    "default": "",
                    "type": "string"
                  }
                },
                "required": [
                  "status"
                ]
              },
              "examples": {
                "default": {
                  "value": {
                    "status": "COMPLETED",
                    "summary": "Fixed the webhook retry and added a regression test.",
                    "key": "pix-payment-bug"
                  },
                  "summary": "Example"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "threadId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "019e4d24-6524-7041-9e1c-8108180cddae"
          },
          {
            "in": "path",
            "name": "issueId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "019e4d24-6524-7041-9e1c-8108180cddaf"
          }
        ],
        "responses": {
          "200": {
            "description": "Declare the lifecycle status of an issue (done / needs input)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "issueId": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "status": {
                      "$ref": "#/components/schemas/IssueStatus"
                    }
                  },
                  "required": [
                    "issueId",
                    "status"
                  ],
                  "additionalProperties": false,
                  "examples": [
                    {
                      "issueId": "019e4d24-6524-7041-9e1c-8108180cddaf",
                      "status": "COMPLETED"
                    }
                  ]
                },
                "examples": {
                  "default": {
                    "value": {
                      "issueId": "019e4d24-6524-7041-9e1c-8108180cddaf",
                      "status": "COMPLETED"
                    },
                    "summary": "Example"
                  }
                }
              }
            }
          }
        },
        "x-mcp-scope": [
          "issue-handling"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StopKind": {
        "type": "string",
        "enum": [
          "SERVER_ERROR",
          "BLOCKED_BY_CLASSIFICATION",
          "HUMAN_REQUESTED",
          "APPROVAL_NEEDED",
          "AUTH_REQUIRED"
        ]
      },
...
      "IssueStatus": {
        "type": "string",
        "enum": [
          "NEEDS_INPUT",
          "WORKING",
          "COMPLETED"
        ]
      },
...
    }
  },
...
}